Receiver trait with target: avoid faulty derefs.
probe.rs follows the chain of Receiver<Target=T> traits for two purposes:
1) to identify potentially relevant `impl` blocks and traits and find
   methods within them;
2) to convert the receiver type to see if it can match the `self` type
   within the methods discovered.

With this change, the chain of `Receiver<Target=T>` is now used
for the first purpose alone. The chain of `Deref<Target=>T` is used for
the second purpose.

Because there's a blanket impl of Receiver for Deref, the Deref chain
is always shorted than the Receiver chain. Therefore we can simply
store a Boolean to represent whether an entry is present in both chains.

From https://github.com/rust-lang/rfcs/pull/3519#discussion_r1400808795