Receiver trait with target: support NotNull Receiver.
This commit allows us to support Receivers which are types that may have
methods implemented upon them in future.

In the event of a conflict between SmartPtr::SomeMethod
and Pointee::Method, favor Pointee::Method because it
was probably added first. Emit a warning (yet to do).

The rationale here is that we want to be able to support

struct Foo;

impl Foo {
  fn method(self: NotNull<Foo>) { }
}

while reserving the right to add future methods to NotNull.
To do that, we have to ensure that any future methods added to NotNull
do not shadow methods added to Foo.

This works by switching some existing hard errors into warnings;
where we previously had two possible method candidates we will
now pick the "innermost" candidate.

This doesn't currently quite work as it produces problems with generics.