Arbitrary self types v2: error on more conflicts.
This version of "arbitrary self types" is the most conservative possible
version. In particular, in the event of any method shadowing, we
intended to produce a hard error. (This masy be relaxed in future).

The probing code already produces an error in this circumstance:

  struct Foo;
  struct SmartPtr<T>(T): // implements Receiver

  impl<T> SmartPtr<T> {
    fn a(self) {}
  }

  impl Foo {
    fn a(self: SmartPtr<Self>) {}
  }

  fn main() {
    let a = SmartPtr(Foo);
    a.a();
  }

but this is because the conflicting methods both take their Self type by
value. The same would apply if they both take the methods by reference,
etc. - but the existing code cannot spot conflicts where one of the
conflicting methods is by value, and the other is by reference
("autoreffed"). This commit enhances the algorithm to spot these
conflicts too.

At present it's not quite good enough, for these reasons:

* It does extra probing and this may be time consuming - not yet
  measured.
* It causes a const generics test to fail, which is just bad luck - the
  extra probing triggers part of const generics which is not yet
  implemented.
* Currently the comparison of the "kind" of each candidate may not be
  exactly right.

A more significant restructuring of probe.rs, with probably roughly the
same semantics, should solve all of these things.