Arbitrary self types v2: pointers feature gate.
The main 'arbitrary_self_types' feature gate will shortly be reused for the
version of arbitrary self type which we intend to stabilize imminently. That's
in some ways a subset of the current feature gate functionality, which supports
raw pointers as receivers. Add a new feature gate for that extra functionality,
which we may come to stabilize in future.