Arbitrary self types v2: Rename old Receiver trait
We're about to introduce a new Receiver trait. There's already a Receiver trait
used to hard-code support for certain smart pointer types whenever the
arbitrary self types feature is _not_ enabled. Rename that old trait to get it
out the way: it will be removed assuming that Arbitrary Self Types becomes
stabilized.

Ideally, we'd keep the same trait and use it in both cases, but this is
unfortunately impossible because stdlib code can't depend on whether some
features are enabled or not, and this would mean that the (old) Receiver trait
was enabled for all types implementing Deref.